fix:csv: archive only when new transactions are imported
Before, the archive rule fired on every import as long as the source
produced any CSV content - even when the .latest mechanism would drop
every row as already-seen. With the data-generating-command form of
source (eg `source | paypalcsv ...`) the source file is never consumed,
so this created a new dated archive file on every run.

The archive trigger now consults previousLatestDates on the rules file
and skips archiving when the filtered journal would be empty.

This required making the .latest helpers accessible to RulesReader,
which Hledger.Read sits above. Move LatestDates, previousLatestDates,
journalFilterSinceLatestDates, etc. to a new Hledger.Read.LatestDates
module; Hledger.Read re-exports them for backward compatibility.
